refactor: split windows spawn resolution from pty
